requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
astroid==3.3.5
asttokens==2.4.1
attrs==21.2.0
bcrypt==3.2.0
beniget==0.4.1
black==24.10.0
blinker==1.4
bloom==0.12.0
breezy==3.2.1
Brlapi==0.8.3
Brotli==1.0.9
builtin-interfaces==1.2.1
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
catkin-tools==0.9.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.18.3
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
comm==0.2.2
composition-interfaces==1.2.1
configobj==5.0.6
cov-core==1.15.0
coverage==6.2
cpuset==1.6
cryptography==3.4.8
cupshelpers==1.0
cv-bridge==3.2.1
cycler==0.11.0
Cython==3.0.11
dbus-python==1.2.18
debugpy==1.8.8
decorator==4.4.2
defer==1.0.6
demo-nodes-py==0.20.5
Deprecated==1.2.13
diagnostic-msgs==4.2.4
diagnostic-updater==4.0.0
dill==0.3.9
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.10.0
dulwich==0.20.31
duplicity==0.8.21
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
exceptiongroup==1.2.2
executing==2.1.0
fastbencode==0.0.5
fasteners==0.14.1
fastimport==0.9.14
ffmpeg-image-transport-msgs==1.1.2
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
geographic-msgs==1.0.6
geometry-msgs==4.2.4
gpg==1.16.0
graphsurgeon==0.4.6
httplib2==0.20.2
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
iniconfig==1.1.1
interactive-markers==2.3.2
ipykernel==6.29.5
ipython==8.29.0
isaac-ros-apriltag-interfaces==3.1.0
isaac-ros-bi3d-interfaces==3.1.0
isaac-ros-launch-utils==0.0.1
isaac-ros-nitros-bridge-interfaces==3.1.0
isaac-ros-nova-interfaces==3.1.0
isaac-ros-pointcloud-interfaces==3.1.0
isaac-ros-tensor-list-interfaces==3.1.0
isaac-ros-test==0.9.0
isort==5.13.2
jedi==0.19.2
jeepney==0.7.1
Jetson.GPIO==2.1.7
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.1.1
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.1
lockfile==0.12.2
logging-demo==0.20.5
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
map-msgs==2.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.7
mccabe==0.6.1
mercurial==6.1.1
message-filters==4.3.5
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
mypy-extensions==1.0.0
nav-msgs==4.2.4
nest-asyncio==1.6.0
netifaces==0.11.0
nmea-msgs==2.0.0
nose2==0.9.2
notify2==0.3
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
onnx-graphsurgeon==0.3.12
osrf-pycommon==2.0.2
packaging==24.2
pandas==1.3.5
paramiko==2.9.3
parso==0.8.4
pathspec==0.12.1
patiencediff==0.2.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
pluggy==0.13.0
ply==3.11
point-cloud-interfaces==1.0.11
prompt_toolkit==3.0.48
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.3.1
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
python-gitlab==2.10.1
python-qt-binding==1.1.2
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzed @ file:///tmp/selfgz6893/pyzed-4.2-cp310-cp310-linux_aarch64.whl
pyzmq==26.2.0
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
realsense2-camera-msgs==4.51.1
requests==2.25.1
requests-toolbelt==0.9.1
resource-retriever==3.1.2
rmw-dds-common==1.6.0
robot-localization==3.5.3
roman==3.3
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rospkg-modules==1.5.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scipy==1.8.0
SecretStorage==3.3.1
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
six==1.16.0
snowballstemmer==2.2.0
sros2==0.10.5
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.9
systemd-python==234
teleop-twist-keyboard==2.4.0
tensorrt==8.6.2
tensorrt-dispatch==8.6.2
tensorrt-lean==8.6.2
tf2-geometry-msgs==0.25.8
tf2-kdl==0.25.8
tf2-msgs==0.25.8
tf2-py==0.25.8
tf2-ros-py==0.25.8
tf2-tools==0.25.8
theora-image-transport==2.5.2
toml==0.10.2
tomli==2.1.0
tomlkit==0.13.2
topic-monitor==0.20.5
tornado==6.4.1
traitlets==5.14.3
trajectory-msgs==4.2.4
turtlesim==1.4.2
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uff==0.6.9
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urllib3==1.26.5
urwid==2.1.2
vcstool==0.3.0
vcstools==0.1.42
visualization-msgs==4.2.4
wadllib==1.3.6
wcwidth==0.2.13
wrapt==1.13.3
xacro==2.0.8
xdg==5
xkit==0.0.0
zed-interfaces==4.1.0
zipp==1.0.0
